Create stylustheme.css
I personally don't like the expanding gamecard thing but as a compromise, I made an addon (Idk what to call it) that overrides the static gamecards and replaces it with your code (it doesn't actually replace it but it has the same effect). I wanted to do the thing that Userstyles have on their website where you can have settings on the style but that's too confusing.  
Sorry for being late to respond but I had some tests that I had to worry about.
The addon: https://github.com/anthony1x6000/ROBLOX2016stylus/blob/downloads/.user.css-installFiles/addons/expandingGameCards.user.css